pre Job done

+++ src/main/java/ru/kata/spring/boot_security/demo/controllers/UsersController.java
@@ -4,6 +4,7 @@ import javax.persistence.EntityNotFoundException;
 import javax.validation.Valid;
 
 import lombok.RequiredArgsConstructor;
+import org.springframework.orm.jpa.JpaObjectRetrievalFailureException;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
@@ -17,11 +18,12 @@ import ru.kata.spring.boot_security.demo.service.UserService;
 import java.util.Collection;
 import java.util.Collections;
 import java.util.List;
+import java.util.Optional;
 
 
 @Controller
 @RequiredArgsConstructor
-@RequestMapping("/users")
+@RequestMapping("/admin")
 public class UsersController {
 
     private final UserService userService;
@@ -38,20 +40,19 @@ public class UsersController {
 
     @GetMapping("/user")
     public String showUser(@RequestParam(value = "id", required = false) Integer id, Model model) {
-
         if (id == null || id <= 0) {
             model.addAttribute("errorMessage", "Ошибка отображения пользователя: значение id не может быть пустым, 0 или отрицательным.");
             return "users/users";
-        } else {
-            try {
-                model.addAttribute("user", userService.showUserById(id));
-            } catch (Exception e) {
-                model.addAttribute("errorMessage", String.format("Ошибка отображения пользователя: %s", e.getMessage()));
-                return "users/users";
-            }
         }
 
-        return "users/user";
+        Optional<User> userOpt = userService.showUserById(id);
+        if (userOpt.isPresent()) {
+            model.addAttribute("user", userOpt.get());
+            return "users/user";
+        } else {
+            model.addAttribute("errorMessage", String.format("Ошибка отображения пользователя: пользователь с ID %d не найден.", id));
+            return "users/users";
+        }
     }
 
     @GetMapping("/new")
@@ -81,7 +82,7 @@ public class UsersController {
         userService.addUser(user);
         redirectAttributes.addFlashAttribute("successMessage",
                 String.format("Пользователь %s успешно создан.", user.getName()));
-        return "redirect:/users";
+        return "redirect:/admin";
     }
 
     @GetMapping("/edit")
@@ -89,17 +90,18 @@ public class UsersController {
         if (id == null || id <= 0) {
             model.addAttribute("errorMessage", "Ошибка редактирования пользователя: значение id не может быть пустым, 0 или отрицательным.");
             return "users/users";
-        } else {
-            try {
-                User user = userService.showUserById(id);
-                model.addAttribute("user", user);
-                model.addAttribute("allRoles", roleRepository.findAll());
-            } catch (EntityNotFoundException e) {
-                model.addAttribute("errorMessage", String.format("Ошибка редактирования пользователя: %s", e.getMessage()));
-                return "users/users";
-            }
         }
-        return "users/edit";
+
+        Optional<User> userOpt = userService.showUserById(id);
+        if (userOpt.isPresent()) {
+            User user = userOpt.get();
+            model.addAttribute("user", user);
+            model.addAttribute("allRoles", roleRepository.findAll());
+            return "users/edit";
+        } else {
+            model.addAttribute("errorMessage", String.format("Ошибка редактирования пользователя: пользователь с ID %d не найден.", id));
+            return "users/users";
+        }
     }
 
     @PostMapping("/update")
@@ -123,27 +125,26 @@ public class UsersController {
         userService.updateUser(user);
         redirectAttributes.addFlashAttribute("successMessage", "Пользователь успешно обновлён.");
 
-        return "redirect:/users/user?id=" + user.getId();
+        return "redirect:/admin/user?id=" + user.getId();
     }
 
 
     @PostMapping("/delete")
     public String deleteUser(@RequestParam(value = "id", required = false) Integer id, RedirectAttributes redirectAttributes) {
-        try {
-            String userName = userService.showUserById(id).getName();
-            userService.deleteUserById(id);
-            redirectAttributes.addFlashAttribute("successMessage", String.format("Пользователь %s успешно удалён.", userName));
-        } catch (Exception e) {
-            redirectAttributes.addFlashAttribute("errorMessage", String.format("Ошибка при удалении пользователя: %s", e.getMessage()));
+        if (id == null || id <= 0) {
+            redirectAttributes.addFlashAttribute("errorMessage", "Некорректный id пользователя.");
+            return "redirect:/admin";
         }
 
-//    } catch (EntityNotFoundException e) {
-//        redirectAttributes.addFlashAttribute("errorMessage", String.format("Пользователь с id=%d не найден.", id));
-//    } catch (IllegalArgumentException e) {
-//        redirectAttributes.addFlashAttribute("errorMessage", "Некорректный id пользователя.");
-//    }
-
-        return "redirect:/users";
+        Optional<User> userOpt = userService.showUserById(id);
+        if (userOpt.isPresent()) {
+            String userName = userOpt.get().getName();
+            userService.deleteUserById(id);
+            redirectAttributes.addFlashAttribute("successMessage", String.format("Пользователь %s успешно удалён.", userName));
+        } else {
+            redirectAttributes.addFlashAttribute("errorMessage", String.format("Ошибка при удалении пользователя: пользователь с ID %d не найден.", id));
+        }
+        return "redirect:/admin";
     }
 
 }
